# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IMD Management API",
        "description": "API documentation for IMD (Inisiasi Menyusu Dini) Management System",
        "contact": {
            "email": "admin@imd.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/imds": {
            "get": {
                "tags": [
                    "IMD"
                ],
                "summary": "Get list of IMD records with filters",
                "operationId": "489a40dc47e0c1422f908009fbd9f65a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in patient name, RM number, or staff name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    },
                    {
                        "name": "cara_persalinan",
                        "in": "query",
                        "description": "Filter by delivery method",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "SC",
                                "Spontan"
                            ]
                        }
                    },
                    {
                        "name": "waktu_imd",
                        "in": "query",
                        "description": "Filter by IMD duration",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "15 menit",
                                "30 menit",
                                "45 menit",
                                "60 menit"
                            ]
                        }
                    },
                    {
                        "name": "tanggal_lahir_dari",
                        "in": "query",
                        "description": "Filter by birth date from",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-01-01"
                        }
                    },
                    {
                        "name": "tanggal_lahir_sampai",
                        "in": "query",
                        "description": "Filter by birth date to",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMD records retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data IMD berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Imd"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "IMD"
                ],
                "summary": "Create new IMD record",
                "operationId": "4e46308f8d440a912878c2e8a2f5783a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImdRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "IMD record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data IMD berhasil ditambahkan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Imd"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/imds/{id}": {
            "get": {
                "tags": [
                    "IMD"
                ],
                "summary": "Get specific IMD record",
                "operationId": "505679981d405f02599c8e3e0f424d49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "IMD ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMD record retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data IMD berhasil diambil"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Imd"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IMD record not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "IMD"
                ],
                "summary": "Update IMD record",
                "operationId": "d018f2ac85514e695e7fa0096b37c745",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "IMD ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImdRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "IMD record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data IMD berhasil diperbarui"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Imd"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IMD record not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "IMD"
                ],
                "summary": "Delete IMD record",
                "operationId": "29111dd2482fcfcbbb43c1d0e19addf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "IMD ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMD record deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data IMD berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IMD record not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Imd": {
                "required": [
                    "id",
                    "nama_pasien",
                    "alamat",
                    "no_rm",
                    "tanggal_lahir",
                    "cara_persalinan",
                    "tanggal_imd",
                    "waktu_imd",
                    "nama_petugas"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01HXYZ123456789"
                    },
                    "nama_pasien": {
                        "type": "string",
                        "example": "Siti Aminah"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123, Jakarta"
                    },
                    "no_rm": {
                        "type": "string",
                        "example": "RM001234"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "cara_persalinan": {
                        "type": "string",
                        "enum": [
                            "SC",
                            "Spontan"
                        ],
                        "example": "Spontan"
                    },
                    "tanggal_imd": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "waktu_imd": {
                        "type": "string",
                        "enum": [
                            "15 menit",
                            "30 menit",
                            "45 menit",
                            "60 menit"
                        ],
                        "example": "30 menit"
                    },
                    "nama_petugas": {
                        "type": "string",
                        "example": "Bidan Sarah"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ImdRequest": {
                "required": [
                    "nama_pasien",
                    "alamat",
                    "no_rm",
                    "tanggal_lahir",
                    "cara_persalinan",
                    "tanggal_imd",
                    "waktu_imd",
                    "nama_petugas"
                ],
                "properties": {
                    "nama_pasien": {
                        "type": "string",
                        "example": "Siti Aminah"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123, Jakarta"
                    },
                    "no_rm": {
                        "type": "string",
                        "example": "RM001234"
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-15"
                    },
                    "cara_persalinan": {
                        "type": "string",
                        "enum": [
                            "SC",
                            "Spontan"
                        ],
                        "example": "Spontan"
                    },
                    "tanggal_imd": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "waktu_imd": {
                        "type": "string",
                        "enum": [
                            "15 menit",
                            "30 menit",
                            "45 menit",
                            "60 menit"
                        ],
                        "example": "30 menit"
                    },
                    "nama_petugas": {
                        "type": "string",
                        "example": "Bidan Sarah"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}